+++ CHANGELOG.md
@@ -24,7 +24,10 @@
 - **app**: added an Etc category
 - **app**: added default image instead of placeholder
 - **app**: added top threads and my friends thread filter
-- **app**: arweave and ipfs support for blur image
+- **app**: added arweave and ipfs support for blur image
+- **app**: added moderation settings to cogno profile
+- **app**: added mini profile to threads
+- **app**: added ability to add and block users
 
 ### Fixed
 
@@ -47,6 +50,7 @@
 - **app**: changed how the image is displayed if no content is present
 - **app**: changed the thread modal width from 3xl to 5xl
 - **app**: changed the error messages to lowercase
+- **app**: changed the font family to monospaced
 
 ### Removed